@@ -22,7 +22,8 @@
     "@eslint/eslintrc": "3.3.1",
     "@eslint/js": "9.26.0",
     "@playwright/test": "1.52.0",
-    "@sveltejs/vite-plugin-svelte": "3.1.2",
+    "@sveltejs/vite-plugin-svelte": "5.0.3",
+    "@sveltejs/vite-plugin-svelte-inspector": "4.0.1",
     "@tsconfig/svelte": "5.0.4",
     "@typescript-eslint/eslint-plugin": "8.32.1",
     "@typescript-eslint/parser": "8.32.1",
@@ -38,12 +39,12 @@
     "postcss-load-config": "6.0.1",
     "prettier": "3.5.3",
     "prettier-plugin-svelte": "3.4.0",
-    "svelte": "4.2.19",
+    "svelte": "5.30.1",
     "svelte-check": "4.2.1",
     "tailwindcss": "3.4.17",
     "tslib": "2.8.1",
     "typescript": "5.8.3",
-    "vite": "5.4.19",
+    "vite": "6.3.5",
     "vite-plugin-css-injected-by-js": "3.5.2",
     "vite-plugin-html": "3.2.2"
   },
