@@ -1,11 +1,22 @@
 <script lang="ts">
 	// SPDX-FileCopyrightText: Magnus Anderssen <magnus@magooweb.com>
 	// SPDX-License-Identifier: AGPL-3.0-or-later
-	export let terms: string;
-	export let original: string;
 
-	let tokenized: { bold: boolean; value: string }[];
-	$: {
+	type Props = {
+		terms: string;
+		original: string;
+	};
+
+	type Token = {
+		bold: boolean;
+		value: string;
+	};
+
+	let { terms, original }: Props = $props();
+
+	let tokenized: Token[] = $state([]);
+
+	$effect.pre(() => {
 		let result = original;
 		terms
 			// remove non simple characters...
@@ -18,7 +29,7 @@
 			bold: token.startsWith('%b%') && token.endsWith('%b%'),
 			value: token.replaceAll(new RegExp('%b%', 'g'), '')
 		}));
-	}
+	});
 </script>
 
 {#each tokenized as token}
