@@ -1,49 +1,61 @@
 <script lang="ts">
 	// SPDX-FileCopyrightText: Magnus Anderssen <magnus@magooweb.com>
 	// SPDX-License-Identifier: AGPL-3.0-or-later
-	import { afterUpdate, onMount } from 'svelte';
+	import { onMount, type Snippet } from 'svelte';
 	import { generateArray } from '../lib/array-utils';
 
 	// eslint-disable-next-line no-undef
 	type T = $$Generic;
+
 	// eslint-disable-next-line @typescript-eslint/no-unused-vars
 	interface $$Slots {
 		default: { item: T };
+		item: { item: T };
 	}
 
-	export let items: T[];
-	export let minCellWidth: number;
-	export let minCellHeight: number;
+	type Props = {
+		items: T[];
+		toto: Snippet<[T]>;
+		minCellWidth: number;
+		minCellHeight: number;
+	};
 
-	const paddingBottom = 10;
+	let { items, minCellHeight, minCellWidth, item }: Props = $props();
 
-	let columns: number;
-	let rows: number;
+	const paddingBottom = 30;
 
-	let me: HTMLDivElement;
-	$: count = items.length;
+	let columns = $state<number | undefined>();
+	let rows = $state<number | undefined>();
+
+	let me = $state<HTMLDivElement | undefined>();
+	let count = $derived(items.length);
 
 	function updateContainer(rows: number, columns: number) {
-		me.style.display = rows === 1 && columns === 1 ? 'block' : 'grid';
-		me.style.overflowY = 'scroll';
-		if (rows > 1) {
-			me.style.gridTemplateRows = generateArray(rows, `${minCellHeight}px`).join(' ');
-		} else {
-			me.style.gridTemplateRows = `${me.getBoundingClientRect().height - paddingBottom}px`;
+		if (me) {
+			me.style.display = rows === 1 && columns === 1 ? 'block' : 'grid';
+			me.style.overflowY = 'scroll';
+			if (rows > 1) {
+				me.style.gridTemplateRows = generateArray(rows, `${minCellHeight}px`).join(' ');
+			} else {
+				me.style.gridTemplateRows = `${me.getBoundingClientRect().height - paddingBottom}px`;
+			}
+			me.style.gridTemplateColumns = generateArray(columns, '1fr').join(' ');
 		}
-		me.style.gridTemplateColumns = generateArray(columns, '1fr').join(' ');
 	}
 
 	function updateCells(rows: number) {
-		const cells = Array.from(me.getElementsByClassName('mwb-dynamic-grid-cell'));
-		cells.forEach((cell) => {
-			const hElement = cell as HTMLElement;
-			if (rows === 1) {
-				hElement.style.height = `${me.getBoundingClientRect().height - paddingBottom}px`;
-			} else {
-				hElement.style.height = 'auto';
+		if (me) {
+			const cells = Array.from(me.getElementsByClassName('mwb-dynamic-grid-cell'));
+			for (const cell of cells) {
+				const hElement = cell as HTMLElement;
+				if (rows === 1) {
+					const top = me.getBoundingClientRect().top;
+					hElement.style.height = `${window.innerHeight - top - paddingBottom}px`;
+				} else {
+					hElement.style.height = 'auto';
+				}
 			}
-		});
+		}
 	}
 
 	function resize() {
@@ -56,25 +68,26 @@
 		updateCells(rows);
 	}
 
-	afterUpdate(() => {
+	$effect(() => {
 		resize();
 	});
 
 	onMount(() => {
-		resize();
-		me.style.paddingBottom = `${paddingBottom}px`;
-		me.style.gridTemplateColumns = '1fr';
-		window.addEventListener('resize', resize);
+		if (me) {
+			me.style.paddingBottom = `${paddingBottom}px`;
+			me.style.gridTemplateColumns = '1fr';
+			window.addEventListener('resize', resize);
+		}
 		return () => {
 			window.removeEventListener('resize', resize);
 		};
 	});
 </script>
 
-<div class="mwb-dynamic-grid" bind:this={me} data-columns={columns} data-rows={rows}>
-	{#each items as item}
+<div bind:this={me} class="mwb-dynamic-grid" data-columns={columns} data-rows={rows}>
+	{#each items as i}
 		<div class="mwb-dynamic-grid-cell">
-			<slot {item} />
+			{@render item(i)}
 		</div>
 	{/each}
 </div>
@@ -85,6 +98,7 @@
 		min-height: 0;
 		min-width: 0;
 	}
+
 	.mwb-dynamic-grid-cell {
 		overflow: hidden;
 		min-width: 0;
