@@ -1,51 +1,54 @@
 <script lang="ts">
+	// SPDX-FileCopyrightText: Magnus Anderssen <magnus@magooweb.com>
+	// SPDX-License-Identifier: AGPL-3.0-or-later
+
 	import { translate } from '@nextcloud/l10n';
-	import { createEventDispatcher } from 'svelte';
-	import type { ChangeEventHandler } from 'svelte/elements';
-	import { writable } from 'svelte/store';
-	import type { Provider } from '../../states/providers';
-	import { APP_NAME, PROVIDER_ALL } from '../../constants';
+	import { get, writable } from 'svelte/store';
+	import { APP_NAME } from '../../constants';
+	import availableProviders, { type Provider } from '../../states/availableProviders';
+	import { isAllSelected, providerIds } from '../../states/query';
+	import { onMount } from 'svelte';
 
-	export let providers: Provider[];
-	/**
-	 * Selection is either PROVIDER_ALL or a list of providers
-	 * if it contains PROVIDER_ALL the rest is ignored
-	 */
-	export let selection: string[];
+	let displayedSelection = writable<string[]>(get(providerIds));
+	let allSelected = $state(true);
 
-	const dispatch = createEventDispatcher();
+	function update(available: Provider[], selected: string[]) {
+		allSelected = isAllSelected(selected);
+		displayedSelection.set(isAllSelected(selected) ? available.map(({ id }) => id) : selected);
+	}
 
-	let displayedSelection = writable(
-		selection.includes(PROVIDER_ALL) ? providers.map(({ id }) => id) : selection
-	);
-	let allProvidersChecked = false;
-	displayedSelection.subscribe((value) => {
-		if (value.length === providers.length) {
-			allProvidersChecked = true;
-			dispatch('update', [PROVIDER_ALL]);
-		} else {
-			allProvidersChecked = false;
-			dispatch('update', value);
-		}
+	availableProviders.subscribe((available) => {
+		update(available, get(providerIds));
 	});
 
-	const updateAll: ChangeEventHandler<HTMLInputElement> = (event) => {
-		const target = event.target as HTMLInputElement;
-		if (target.checked) {
-			displayedSelection.set(providers.map(({ id }) => id));
-		} else {
-			displayedSelection.set([]);
+	displayedSelection.subscribe((values) => {
+		const available = get(availableProviders);
+		allSelected = isAllSelected(values);
+		providerIds.set(
+			available.length === values.length || !values.length
+				? available.map(({ id }) => id)
+				: values
+		);
+	});
+
+	function toggleAll() {
+		if (!isAllSelected(get(displayedSelection))) {
+			displayedSelection.set(get(availableProviders).map(({ id }) => id));
 		}
-	};
+	}
+
+	onMount(() => {
+		update(get(availableProviders), get(providerIds));
+	});
 </script>
 
 <div class="mwb-checkboxes-container">
 	<label>
-		<input checked={allProvidersChecked} name="providers" on:change={updateAll} type="checkbox" />
+		<input checked={allSelected} name="providers" onchange={toggleAll} type="checkbox" />
 		<span>{translate(APP_NAME, 'All providers')}</span>
 	</label>
 	<div class="mwb-checkboxes">
-		{#each providers as provider}
+		{#each $availableProviders as provider}
 			<label>
 				<input
 					type="checkbox"
@@ -66,7 +69,7 @@
 	}
 
 	.mwb-checkboxes {
-		@apply flex  flex-wrap gap-x-6;
+		@apply flex flex-wrap gap-x-6;
 	}
 
 	label {
@@ -80,6 +83,7 @@
 			@apply whitespace-nowrap cursor-pointer;
 		}
 	}
+
 	/* mwb-tabbed is set on the app container */
 	:global(.mwb-tabbed) {
 		label:focus-within {
